pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bse-predict"
version = "1.0.0"
description = "Bitcoin Sol Eth Predict - Multi-target crypto price prediction system"
authors = [{name = "BSE Predict Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core ML and Data
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    
    # Database
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    
    # Exchange API
    "ccxt>=4.2.0",
    
    # Telegram
    "python-telegram-bot>=20.8",
    
    # Utilities
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "flask>=3.0.0",
    "psutil>=5.9.8",
    
    # Development
    "pytest>=8.0.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
